EM-3401 Rename TestConstants as its purpose has evolved
