+++ src/integrationTest/java/uk/gov/hmcts/reform/em/hrs/ingestor/av/AntivirusClientIntegrationTest.java
@@ -5,7 +5,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ContextConfiguration;
 import uk.gov.hmcts.reform.em.hrs.ingestor.av.mock.ClamAvInitializer;
 import uk.gov.hmcts.reform.em.hrs.ingestor.config.ClamAvConfig;
-import uk.gov.hmcts.reform.em.hrs.ingestor.helper.TestConstants;
+import uk.gov.hmcts.reform.em.hrs.ingestor.helper.TestUtil;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -14,8 +14,8 @@ import java.net.URISyntaxException;
 import javax.inject.Inject;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.hmcts.reform.em.hrs.ingestor.helper.TestConstants.CLEAN_FILE;
-import static uk.gov.hmcts.reform.em.hrs.ingestor.helper.TestConstants.INFECTED_FILE;
+import static uk.gov.hmcts.reform.em.hrs.ingestor.helper.TestUtil.CLEAN_FILE;
+import static uk.gov.hmcts.reform.em.hrs.ingestor.helper.TestUtil.INFECTED_FILE;
 
 @SpringBootTest(classes = {ClamAvConfig.class, AntivirusClientImpl.class})
 @ContextConfiguration(initializers = {ClamAvInitializer.class})
@@ -45,7 +45,7 @@ class AntivirusClientIntegrationTest {
     }
 
     private InputStream getFileAsInputStream(final String filename) throws IOException, URISyntaxException {
-        final byte[] data = TestConstants.getFileContent(filename);
+        final byte[] data = TestUtil.getFileContent(filename);
         return new ByteArrayInputStream(data);
     }
