+++ src/integrationTest/java/uk/gov/hmcts/reform/em/hrs/ingestor/http/HrsApiClientIntegrationTest.java
@@ -26,7 +26,7 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIOException;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static uk.gov.hmcts.reform.em.hrs.ingestor.helper.TestConstants.convertObjectToJsonString;
+import static uk.gov.hmcts.reform.em.hrs.ingestor.helper.TestUtil.convertObjectToJsonString;
 
 @SpringBootTest(classes = {TestOkHttpClientConfig.class, AppConfig.class, HrsApiClientImpl.class})
 @ContextConfiguration(initializers = {WireMockInitializer.class})
